cdk/application/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via starlette
boto3==1.34.91
    # via -r requirements.in
botocore==1.34.91
    # via
    #   boto3
    #   s3transfer
exceptiongroup==1.2.1
    # via anyio
fastapi==0.110.2
    # via -r requirements.in
idna==3.7
    # via anyio
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
mangum==0.17.0
    # via -r requirements.in
pydantic==2.7.1
    # via fastapi
pydantic-core==2.18.2
    # via pydantic
pyjwt==2.8.0
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via botocore
s3transfer==0.10.1
    # via boto3
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
starlette==0.37.2
    # via fastapi
typing-extensions==4.11.0
    # via
    #   anyio
    #   fastapi
    #   mangum
    #   pydantic
    #   pydantic-core
    #   starlette
urllib3==1.26.18
    # via botocore
